% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data}
\alias{data}
\title{Bozeman Bike Commuter Data}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(bozemanbike)
}
\description{
Data from a Bozeman Bike Commuter Counts
}
\examples{
data(bozemanbike)
}
\keyword{datasets}
