% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_function.R
\name{datagen}
\alias{datagen}
\title{A Spatial Process Function for Bozeman Bike Commuter Data}
\usage{
datagen(sigma.sq, tau.sq, phi, beta)
}
\arguments{
\item{sigma.sq}{You know what it is}

\item{tau.sq}{}

\item{phi}{}

\item{beta}{covariate slope}
}
\description{
why am i so bad at this...
}
\examples{
datagen()
}
\keyword{Bozeman}
