% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save.data}
\alias{save.data}
\title{A Spatial Process Function for Bozeman Bike Commuter Data}
\usage{
save.data()
}
\arguments{
\item{none}{no covariates here}
}
\description{
This function is so ch1LL and makes fake bike count data for our lovely city
}
\examples{
save.data()

}
\keyword{Bozeman}
