% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_function.R
\name{datagen}
\alias{datagen}
\title{A Spatial Process Function for Bozeman Bike Commuter Data}
\usage{
datagen(sigma.sq, tau.sq, phi)
}
\arguments{
\item{sigma.sq}{You know what it is}

\item{tau.sq}{You know what it is}

\item{phi}{You know what it is}
}
\description{
This function is so ch1LL and makes fake bike count data for our lovely city
}
\examples{
datagen()
}
\keyword{Bozeman}
